+++ Sources/WhisperKit/MLX/MLXModels.swift
@@ -1,0 +1,9 @@
+//  For licensing see accompanying LICENSE.md file.
+//  Copyright © 2024 Argmax, Inc. All rights reserved.
+
+import Foundation
+
+public enum PadMode {
+    case constant
+    case reflect
+}
